Update Remote Command for JSON
