+++ Cartfile.resolved
@@ -1,3 +1,2 @@
-binary "https://tags.tiqcdn.com/dle/tealiummobile/tealium-ios-carthage/tealium-carthage-plcrashreporter.json" "1.3.6"
-github "ContentSquare/CS_iOS_SDK" "v3.0.0"
-github "tealium/tealium-swift" "1.9.1"
+git "file:///Users/christina/Projects/work-repos/github/tealium-swift-builder" "388d0ac6726fd110a03c48ad14cdfa8ae87706e1"
+github "ContentSquare/CS_iOS_SDK" "v3.2.1"
