+++ Sources/ContentsquareRemoteCommand.swift
@@ -3,7 +3,7 @@
 //  TealiumContentsquare
 //
 //  Created by Jonathan Wong on 3/6/20.
-//  Copyright © 2020 Jonathan Wong. All rights reserved.
+//  Copyright © 2020 Tealium. All rights reserved.
 //
 
 import Foundation
@@ -15,42 +15,42 @@ import TealiumTagManagement
 import TealiumRemoteCommands
 #endif
 
-public class ContentsquareRemoteCommand {
+public class ContentsquareRemoteCommand: RemoteCommand {
     
-    let contentsquareTracker:
-    ContentsquareTrackable
+    var contentsquareTracker: ContentsquareTrackable?
     
-    public init(contentsquareTracker:
-        ContentsquareTrackable) {
-        self.contentsquareTracker =
-        contentsquareTracker
+    public init(contentsquareTracker: ContentsquareTrackable = ContentsquareTracker(), type: RemoteCommandType = .webview) {
+        self.contentsquareTracker = contentsquareTracker
+        weak var selfWorkaround: ContentsquareRemoteCommand?
+        super.init(commandId: Contentsquare.commandId,
+                   description: Contentsquare.description,
+            type: type,
+            completion: { response in
+                guard let payload = response.payload else {
+                    return
+                }
+                selfWorkaround?.processRemoteCommand(with: payload)
+            })
+        selfWorkaround = self
     }
     
-    public func remoteCommand() -> TealiumRemoteCommand {
-        return TealiumRemoteCommand(commandId: "contentsquare",
-                                    description: "Contentsquare Remote Command") { response in
-                                        let payload = response.payload()
-                                        guard let command = payload[Contentsquare.Commands.commandKey] as? String else {
-                                            return
-                                        }
-                                        
-                                        let commands = command.split(separator: Contentsquare.Commands.separator)
-                                        let formatted = commands.map { command in
-                                            return command.trimmingCharacters(in: CharacterSet.whitespacesAndNewlines)
-                                        }
-                                        
-                                        self.parseCommands(formatted, payload: payload)
+    func processRemoteCommand(with payload: [String: Any]) {
+        guard let contentsquareTracker = contentsquareTracker,
+              let command = payload[Contentsquare.commandKey] as? String else {
+                return
         }
-    }
-    
-    func parseCommands(_ commands: [String], payload: [String: Any]) {
-        commands.forEach { command in
-            let lowercasedCommand = command.lowercased()
-            switch lowercasedCommand {
-            case Contentsquare.Commands.sendScreenView:
+        let commands = command.split(separator: Contentsquare.separator)
+        let contentsquareCommands = commands.map { command in
+            return command.trimmingCharacters(in: CharacterSet.whitespacesAndNewlines)
+        }
+
+        contentsquareCommands.forEach {
+            let command = Contentsquare.Commands(rawValue: $0.lowercased())
+            switch command {
+            case .sendScreenView:
                 guard let screenName = payload[Contentsquare.ScreenView.screenName] as? String else { return }
                 contentsquareTracker.sendScreenView(screenName: screenName)
-            case Contentsquare.Commands.sendTransaction:
+            case .sendTransaction:
                 guard let options = payload[Contentsquare.TransactionProperties.transaction] as? [String: Any] else {
                     print("Contentsquare.TransactionProperties.transaction key is missing.")
                     return
@@ -59,18 +59,18 @@ public class ContentsquareRemoteCommand {
                     let currency: String = options[Contentsquare.TransactionProperties.currency] as? String else { return }
                 let transactionId: String? = options[Contentsquare.TransactionProperties.transactionId] as? String
                 contentsquareTracker.sendTransaction(price: price, currency: currency, transactionId: transactionId)
-            case Contentsquare.Commands.sendDynamicVar:
+            case .sendDynamicVar:
                 guard let dynamicVar = payload[Contentsquare.DynamicVar.dynamicVar] as? [String: Any] else { return }
                 contentsquareTracker.sendDynamicVar(dynamicVar: dynamicVar)
-            case Contentsquare.Commands.stopTracking:
+            case .stopTracking:
                 contentsquareTracker.stopTracking()
-            case Contentsquare.Commands.resumeTracking:
+            case .resumeTracking:
                 contentsquareTracker.resumeTracking()
-            case Contentsquare.Commands.forgetMe:
+            case .forgetMe:
                 contentsquareTracker.forgetMe()
-            case Contentsquare.Commands.optIn:
+            case .optIn:
                 contentsquareTracker.optIn()
-            case Contentsquare.Commands.optOut:
+            case .optOut:
                 contentsquareTracker.optOut()
             default: break
             }
