+++ TealiumContentsquare.podspec
@@ -3,7 +3,7 @@ Pod::Spec.new do |s|
     # ―――  Spec Metadata  ―――――――――――――――――――――――――――――――――――――――――――――――――――――――――― #
     s.name         = "TealiumContentsquare"
     s.module_name  = "TealiumContentsquare"
-    s.version      = "1.0.0"
+    s.version      = "2.0.0"
     s.summary      = "Tealium Swift and Contentsquare integration"
     s.description  = <<-DESC
     Tealium's integration with Contentsquare for iOS.
@@ -31,10 +31,9 @@ Pod::Spec.new do |s|
     s.ios.source_files      = "Sources/*.{swift}"
 
     # ――― Dependencies ―――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――― #
-    s.ios.dependency 'tealium-swift/Core'
-    s.ios.dependency 'tealium-swift/TealiumRemoteCommands'
-    s.ios.dependency 'tealium-swift/TealiumDelegate'
-    s.ios.dependency 'tealium-swift/TealiumTagManagement'
+    s.ios.dependency 'tealium-swift/Core', '~> 2.1'
+    s.ios.dependency 'tealium-swift/RemoteCommands', '~> 2.1'
+    s.ios.dependency 'tealium-swift/TagManagement', '~> 2.1'
     s.ios.dependency 'CS_iOS_SDK'
 
 end
