+++ TealiumContentsquareExample/TealiumContentsquareExample/TealiumHelper.swift
@@ -10,44 +10,55 @@ import Foundation
 import TealiumSwift
 import TealiumContentsquare
 
+enum TealiumConfiguration {
+    static let account = "tealiummobile"
+    static let profile = "contentsquare-dev"
+    static let environment = "qa"
+}
+
 class TealiumHelper {
-    
+
     static let shared = TealiumHelper()
-    let config = TealiumConfig(account: "tealiummobile",
-                               profile: "contentsquare-dev",
-                               environment: "qa",
-                               datasource: nil)
-    
+
+    let config = TealiumConfig(account: TealiumConfiguration.account,
+        profile: TealiumConfiguration.profile,
+        environment: TealiumConfiguration.environment)
+
     var tealium: Tealium?
+    
     static var universalData = [String: Any]()
     
+    // JSON Remote Command
+    let contentsquareRemoteCommand = ContentsquareRemoteCommand(type: .remote(url: "https://tags.tiqcdn.com/dle/tealiummobile/demo/contentsquare.json"))
+
     private init() {
-        let list = TealiumModulesList(isWhitelist: false,
-                                      moduleNames: ["autotracking", "collect", "consentmanager"])
-        config.modulesList = list
-        config.logLevel = .verbose
-        config.shouldUseRemotePublishSettings = true
+        config.shouldUseRemotePublishSettings = false
         config.batchingEnabled = false
+        config.remoteAPIEnabled = true
+        config.logLevel = .info
+        config.dispatchers = [Dispatchers.TagManagement, Dispatchers.RemoteCommands]
         
-        tealium = Tealium(config: config) { responses in
-            guard let remoteCommands = self.tealium?.remoteCommands() else {
-                return
-            }
-            let contentsquareTracker = ContentsquareTracker()
-            let contentsquareCommand = ContentsquareRemoteCommand(contentsquareTracker: contentsquareTracker)
-            let contentsquareRemoteCommand = contentsquareCommand.remoteCommand()
-            remoteCommands.add(contentsquareRemoteCommand)
-        }
+        config.addRemoteCommand(contentsquareRemoteCommand)
+        
+        tealium = Tealium(config: config)
     }
-    
+
+
     class func start() {
         _ = TealiumHelper.shared
     }
-    
+
+    class func trackView(title: String, data: [String: Any]?) {
+        let tealiumView = TealiumView(title, dataLayer: data)
+        TealiumHelper.shared.tealium?.track(tealiumView)
+    }
+
     class func track(title: String, data: [String: Any]?) {
         if let data = data {
             universalData = universalData.merging(data) { _, new in new }
         }
-        TealiumHelper.shared.tealium?.track(title: title, data: universalData, completion: nil)
+        let tealiumEvent = TealiumEvent(title, dataLayer: universalData)
+        TealiumHelper.shared.tealium?.track(tealiumEvent)
     }
+
 }
