+++ Tests/ContentsquareTrackerTests.swift
@@ -17,66 +17,60 @@ class ContentsquareTrackerTests: XCTestCase {
 
     let contentsquareTracker = MockContentsquareTracker()
     var contentsquareCommand: ContentsquareRemoteCommand!
-    var remoteCommand: TealiumRemoteCommand!
     
     override func setUp() {
         contentsquareCommand = ContentsquareRemoteCommand(contentsquareTracker: contentsquareTracker)
-        remoteCommand = contentsquareCommand.remoteCommand()
-    }
-
-    override func tearDown() {
-        
     }
 
     func testScreenViewCalledWithKey() {
-        contentsquareCommand.parseCommands(["sendscreenview"], payload: ["screen_name": "home"])
+        contentsquareCommand.processRemoteCommand(with: ["command_name": "sendscreenview", "screen_name": "home"])
         XCTAssertEqual(1, contentsquareTracker.sendScreenViewCallCount)
     }
 
     func testScreenViewNotCalledWitouthKey() {
-        contentsquareCommand.parseCommands(["sendscreenview"], payload: ["not_screen_name": "home"])
+        contentsquareCommand.processRemoteCommand(with: ["command_name": "sendscreenview", "not_screen_name": "home"])
         XCTAssertEqual(0, contentsquareTracker.sendScreenViewCallCount)
     }
 
     func testTransactionCalledWithKey() {
-        contentsquareCommand.parseCommands(["sendtransaction"], payload: ["transaction": ["price": 1.99,
+        contentsquareCommand.processRemoteCommand(with: ["command_name": "sendtransaction", "transaction": ["price": 1.99,
                                                                                           "currency": "USD"]])
         XCTAssertEqual(1, contentsquareTracker.sendTransactionCallCount)
     }
     
     func testTransactionNotCalledWithoutKey() {
-        contentsquareCommand.parseCommands(["sendtransaction"], payload: ["not_transaction": ["price": 1.99,
+        contentsquareCommand.processRemoteCommand(with: ["command_name": "sendtransaction", "not_transaction": ["price": 1.99,
                                                                                           "currency": Int(1)]])
         XCTAssertEqual(0, contentsquareTracker.sendTransactionCallCount)
     }
     
     func testTransactionCalledWithIdKey() {
-        contentsquareCommand.parseCommands(["sendtransaction"], payload: ["transaction": ["price": 1.99,
+        contentsquareCommand.processRemoteCommand(with: ["command_name": "sendtransaction", "transaction": ["price": 1.99,
                                                                                           "currency": "CAD",
                                                                                           "id": "123"]])
         XCTAssertEqual(1, contentsquareTracker.sendTransactionCallCount)
     }
     
     func testTransactionNotCalledWithoutPriceKey() {
-        contentsquareCommand.parseCommands(["sendtransaction"], payload: ["not_transaction": ["not_price": 1.99,
+        contentsquareCommand.processRemoteCommand(with: ["command_name": "sendtransaction", "not_transaction": ["not_price": 1.99,
                                                                                               "currency": Int(1)]])
         XCTAssertEqual(0, contentsquareTracker.sendTransactionCallCount)
     }
     
     func testTransactionNotCalledWithoutCurrencyKey() {
-        contentsquareCommand.parseCommands(["sendtransaction"], payload: ["not_transaction": ["price": 1.99,
+        contentsquareCommand.processRemoteCommand(with: ["command_name": "sendtransaction", "not_transaction": ["price": 1.99,
                                                                                               "not_currency": Int(1)]])
         XCTAssertEqual(0, contentsquareTracker.sendTransactionCallCount)
     }
     
     func testDynamicVarCalledWithKey() {
-        contentsquareCommand.parseCommands(["senddynamicvar"], payload: ["dynamic_var": ["not_price": 1.99,
+        contentsquareCommand.processRemoteCommand(with: ["command_name": "senddynamicvar", "dynamic_var": ["not_price": 1.99,
                                                                                               "currency": Int(1)]])
         XCTAssertEqual(0, contentsquareTracker.sendTransactionCallCount)
     }
     
     func testDynamicVarNotCalledWithoutKey() {
-        contentsquareCommand.parseCommands(["senddynamicvar"], payload: ["not_dynamic_var": ["price": 1.99,
+        contentsquareCommand.processRemoteCommand(with: ["command_name": "senddynamicvar", "not_dynamic_var": ["price": 1.99,
                                                                                               "not_currency": Int(1)]])
         XCTAssertEqual(0, contentsquareTracker.sendTransactionCallCount)
     }
